Refactored the staking pallet
