+++ runtime/src/configs/mod.rs
@@ -1,9 +1,17 @@
 pub mod authorship;
-pub mod balances;
 pub mod bags_list;
+pub mod balances;
+// pub mod collective;
 pub mod timestamp;
 pub mod transaction_payment;
+
+pub mod election_provider_multi_phase;
+pub use election_provider_multi_phase::*;
+
 pub mod membership;
 pub mod im_online;
 pub mod offences;
 pub mod session;
+
+pub mod staking;
+pub use staking::*;
